cpdetect: refactor
