+++ src/cpdetect/heuristics/heuristics.cpp
@@ -1,5 +1,5 @@
 /**
- * @file src/cpdetect/compiler_detector/heuristics/heuristics.cpp
+ * @file src/cpdetect/heuristics/heuristics.cpp
  * @brief Class for heuristics detection.
  * @copyright (c) 2017 Avast Software, licensed under the MIT license
  */
@@ -16,7 +16,7 @@
 #include "retdec/utils/container.h"
 #include "retdec/utils/conversion.h"
 #include "retdec/utils/string.h"
-#include "retdec/cpdetect/compiler_detector/heuristics/heuristics.h"
+#include "retdec/cpdetect/heuristics/heuristics.h"
 #include "retdec/fileformat/utils/conversions.h"
 
 using namespace retdec::utils;
