+++ src/fileinfo/file_detector/coff_detector.cpp
@@ -525,7 +525,7 @@ void CoffDetector::getAdditionalInfo()
  */
 retdec::cpdetect::CompilerDetector* CoffDetector::createCompilerDetector() const
 {
-	return new CoffCompiler(*coffParser, cpParams, fileInfo.toolInfo);
+	return new CompilerDetector(*coffParser, cpParams, fileInfo.toolInfo);
 }
 
 } // namespace fileinfo
