+++ src/fileinfo/file_detector/elf_detector.cpp
@@ -1969,7 +1969,7 @@ void ElfDetector::getAdditionalInfo()
  */
 retdec::cpdetect::CompilerDetector* ElfDetector::createCompilerDetector() const
 {
-	return new ElfCompiler(*elfParser, cpParams, fileInfo.toolInfo);
+	return new CompilerDetector(*elfParser, cpParams, fileInfo.toolInfo);
 }
 
 } // namespace fileinfo
