+++ src/fileinfo/file_detector/intel_hex_detector.cpp
@@ -138,7 +138,7 @@ void IntelHexDetector::getAdditionalInfo()
  */
 retdec::cpdetect::CompilerDetector* IntelHexDetector::createCompilerDetector() const
 {
-	return new IntelHexCompiler(*ihexParser, cpParams, fileInfo.toolInfo);
+	return new CompilerDetector(*ihexParser, cpParams, fileInfo.toolInfo);
 }
 
 } // namespace fileinfo
