+++ src/fileinfo/file_detector/intel_hex_detector.h
@@ -8,6 +8,7 @@
 #define FILEINFO_FILE_DETECTOR_INTEL_HEX_DETECTOR_H
 
 #include "fileinfo/file_detector/file_detector.h"
+#include "retdec/fileformat/file_format/intel_hex/intel_hex_format.h"
 
 namespace retdec {
 namespace fileinfo {
