+++ src/fileinfo/file_detector/raw_data_detector.h
@@ -8,6 +8,7 @@
 #define FILEINFO_FILE_DETECTOR_RAW_DATA_DETECTOR_H
 
 #include "fileinfo/file_detector/file_detector.h"
+#include "retdec/fileformat/file_format/raw_data/raw_data_format.h"
 
 namespace retdec {
 namespace fileinfo {
